+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 8.13.5 (Unreleased)
+
+### Miscellaneous
+
+ - Improve dependency install script
+
+
 ## 8.13.4 (2022-04-03)
 
 ### Bugfixes